# shared/api-v1.openAPI.yaml
openapi: 3.0.3
info:
  title: Baziche API
  version: 1.0.0
  description: Phase 2 — auth, projects (+restore/merge), assets (magic-byte verified), meta, admin. builds/billing in later phases.
servers:
  - url: https://api.baziche.app/api/v1
  - url: http://localhost:8787/api/v1
paths:
...
  /projects:
    get:
      summary: List my projects
      security: [ { bearer: [] } ]
      responses: { '200': { description: '{ success, projects[] }' } }
    post:
      summary: Create project
      security: [ { bearer: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [name, gameType], properties: { name: { type: string }, gameType: { type: string } } } } } }
      responses: { '200': { description: '{ success, project }' } }
  /projects/{id}:
    get:
      summary: Get project meta + current JSON
      security: [ { bearer: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: '{ success, project, json }' }, '404': { $ref: '#/components/responses/Error' } }
    patch:
      summary: Save project JSON (baseRev guarded, 409 on conflict)
      security: [ { bearer: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [baseRev, json], properties: { baseRev: { type: integer }, name: { type: string }, json: { type: object } } } } } }
      responses: { '200': { description: '{ success, rev }' }, '409': { description: '{ success:false, error:REVISION_CONFLICT, serverRev, serverCopy }' } }
    delete:
      summary: Soft delete
      security: [ { bearer: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: ok } }
  /projects/{id}/revisions:
    get:
      summary: List revisions
      security: [ { bearer: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: '{ success, revisions[] }' } }
  /projects/{id}/restore:
    post:
      summary: Restore an old revision as the new head (baseRev guarded)
      security: [ { bearer: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [rev, baseRev], properties: { rev: { type: integer }, baseRev: { type: integer } } } } } }
      responses: { '200': { description: '{ success, rev, restoredFrom }' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /projects/{id}/merge:
    post:
      summary: Three-way merge (baseRev + mine vs server; server wins flagged conflicts)
      security: [ { bearer: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [baseRev, json], properties: { baseRev: { type: integer }, json: { type: object } } } } } }
      responses: { '200': { description: '{ success, merged, conflicts[], serverRev }' } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Error:
      description: '{ success:false, error:{ code, message }, meta? }'